Move schedAdherence code
